add coeff entity & new logic

+++ Personal-Testing-System/Models/EmployeeResultSubcompetenceModel.cs
@@ -8,7 +8,7 @@ namespace Personal_Testing_System.Models
         public int? ScoreFrom { get; set; }
         public int? ScoreTo { get; set; }
         public string? ResultLevel { get; set; }
-        public int? NumberPoints { get; set; }
+        public double? NumberPoints { get; set; }
         public List<ResultSubcompetenceModel>? SubcompetenceResults { get; set; }
         public ResultModel? Result { get; set; }
         public EmployeeModel? Employee { get; set; }
